Ensure grunt fails when tests fail
The previous code was just calling grunt async done when the tests
failed. It needs to pass a boolean for whether the task actually failed
or not.

The code returned will be 0 if successful and a number (usually 1) if
not. So we pass the inverse of this to fail or pass the task.